chore: move list template to new file
